+++ Abel/head/OptimizationProblems/lprog_Result.h
@@ -6,8 +6,8 @@ struct lprog_Result {
 	double error;
 	size_t niter;
 	bool isDiverging;
-	int z_num; // number of dual variables for equality constraints
-	int s_num; // number of dual variables for inequalities
+	size_t z_num; // number of dual variables for equality constraints
+	size_t s_num; // number of dual variables for inequalities
 
-	lprog_Result(const Eigen::VectorXd& result, const Eigen::VectorXd& dual_variables, double error, size_t niter, bool isDiverging, int z_num, int s_num);
+	lprog_Result(const Eigen::VectorXd& result, const Eigen::VectorXd& dual_variables, double error, size_t niter, bool isDiverging, size_t z_num, size_t s_num);
 };
