+++ Abel/src/Base/MVNumericalMethod.cpp
@@ -1,9 +1,9 @@
 #include "pch.h"
 #include "Base/MVNumericalMethod.h"
 
-MVNumericalMethod::MVNumericalMethod(size_t dimension, double tol, int max_iter, bool was_run, int iter_counter, double err, 
+MVNumericalMethod::MVNumericalMethod(size_t dimension, double tol, size_t max_iter, bool was_run, int iter_counter, double err,
 	const Eigen::VectorXd& res, const Eigen::VectorXd& starting_point, bool hasStart) :
-	tol(std::min(std::max(1e-15, tol), 0.1)), max_iter(std::max(2, max_iter)), iter_counter(iter_counter), was_run(was_run), error(err), result(res), dimension(dimension), 
+	tol(std::min(std::max(1e-15, tol), 0.1)), max_iter(std::max((size_t)2, max_iter)), iter_counter(iter_counter), was_run(was_run), error(err), result(res), dimension(dimension),
 	starting_point(starting_point), hasStart(hasStart)
 {
 	if (result.size() == 0) {
